requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==6.0.0          # via graphene
django-environ==0.4.5
django-extensions==2.1.9
django-filter==2.1.0
django-filters==0.2.1
django-guardian==2.0.0
django==2.2.3
djangorestframework==3.9.4
graphene-django==2.3.2
graphene==2.1.6           # via graphene-django
graphql-core==2.2         # via graphene, graphene-django, graphql-relay
graphql-relay==0.4.5      # via graphene
model-mommy==1.6.0
promise==2.2.1            # via graphene-django, graphql-core, graphql-relay
psycopg2==2.8.3
pytz==2019.1              # via django
rx==1.6.1                 # via graphql-core
singledispatch==3.4.0.3   # via graphene-django
six==1.12.0               # via django-extensions, graphene, graphene-django, graphql-core, graphql-relay, model-mommy, promise, singledispatch
sqlparse==0.3.0           # via django
